Created Dashboard UI of the application
Includes two buttons to view inventory and to log daily usage of ingredients. Additionally has a feature to view the newly changed ingredients in the inventory. This is however hardcoded for now till the backend developer develops her share. Additionally it includes indicators which will show case the amount of new additions to the inventory, low stock and bad quality stock.